migrate `debug,evm,hardhat,net,web3` namespaces to jsonrpsee
