+++ Cargo.toml
@@ -1,8 +1,11 @@
 [workspace]
 members = [
+    "crates/api_decl",
+    "crates/api_server",
     "crates/cli",
     "crates/config",
     "crates/core",
+    "crates/types",
 ]
 resolver = "2"
 
@@ -53,6 +56,7 @@ jsonrpc-http-server = { git = "https://github.com/matter-labs/jsonrpc.git", bran
 jsonrpc-core = { git = "https://github.com/matter-labs/jsonrpc.git", branch = "master" }
 jsonrpc-core-client = { git = "https://github.com/matter-labs/jsonrpc.git", branch = "master" }
 jsonrpc-derive = { git = "https://github.com/matter-labs/jsonrpc.git", branch = "master" }
+jsonrpsee = "0.23"
 lazy_static = "1.4"
 once_cell = "1.7"
 rand = "0.8"
@@ -75,5 +79,8 @@ tracing-subscriber = { version = "0.3", features = [
 #########################
 # Local dependencies    #
 #########################
+anvil_zksync_api_decl = { path = "crates/api_decl" }
+anvil_zksync_api_server = { path = "crates/api_server" }
 anvil_zksync_config = { path = "crates/config" }
 anvil_zksync_core = { path = "crates/core" }
+anvil_zksync_types = { path = "crates/types" }
