+++ crates/api_decl/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "anvil_zksync_api_decl"
+description = "anvil-zksync API declarations"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+keywords.workspace = true
+categories.workspace = true
+
+[dependencies]
+anvil_zksync_types.workspace = true
+
+zksync_types.workspace = true
+zksync_web3_decl.workspace = true
+
+jsonrpsee.workspace = true
