+++ crates/api_decl/src/lib.rs
@@ -1,0 +1,7 @@
+mod namespaces;
+
+pub use namespaces::{AnvilNamespaceServer, EvmNamespaceServer};
+
+pub use zksync_web3_decl::namespaces::{
+    DebugNamespaceServer, NetNamespaceServer, Web3NamespaceServer,
+};
