+++ crates/api_decl/src/namespaces/anvil.rs
@@ -1,13 +1,13 @@
-use super::{ResetRequest, RpcResult};
-use crate::utils::Numeric;
-use jsonrpc_derive::rpc;
-use serde::{Deserialize, Serialize};
-use zksync_types::api::{Block, Transaction};
+use anvil_zksync_types::api::{DetailedTransaction, ResetRequest};
+use anvil_zksync_types::Numeric;
+use jsonrpsee::core::RpcResult;
+use jsonrpsee::proc_macros::rpc;
+use zksync_types::api::Block;
 use zksync_types::web3::Bytes;
 use zksync_types::{Address, H256, U256, U64};
 
-#[rpc]
-pub trait AnvilNamespaceT {
+#[rpc(server, namespace = "anvil")]
+pub trait AnvilNamespace {
     /// Create a buffer that represents all state on the chain, which can be loaded to separate
     /// process by calling `anvil_loadState`.
     ///
@@ -17,7 +17,7 @@ pub trait AnvilNamespaceT {
     ///
     /// # Returns
     /// Buffer representing the chain state.
-    #[rpc(name = "anvil_dumpState")]
+    #[method(name = "dumpState", aliases = ["hardhat_dumpState"])]
     fn dump_state(&self, preserve_historical_states: Option<bool>) -> RpcResult<Bytes>;
 
     /// Append chain state buffer to current chain. Will overwrite any conflicting addresses or
@@ -29,14 +29,14 @@ pub trait AnvilNamespaceT {
     ///
     /// # Returns
     /// `true` if a snapshot was reverted, otherwise `false`.
-    #[rpc(name = "anvil_loadState")]
+    #[method(name = "loadState", aliases = ["hardhat_loadState"])]
     fn load_state(&self, bytes: Bytes) -> RpcResult<bool>;
 
     /// Mines a single block in the same way as `evm_mine` but returns extra fields.
     ///
     /// # Returns
     /// Freshly mined block's representation along with extra fields.
-    #[rpc(name = "anvil_mine_detailed")]
+    #[method(name = "mine_detailed", aliases = ["evm_mine_detailed"])]
     fn mine_detailed(&self) -> RpcResult<Block<DetailedTransaction>>;
 
     /// Sets the fork RPC url. Assumes the underlying chain is the same as before.
@@ -44,7 +44,7 @@ pub trait AnvilNamespaceT {
     /// # Arguments
     ///
     /// * `url` - Fork's new URL
-    #[rpc(name = "anvil_setRpcUrl")]
+    #[method(name = "setRpcUrl")]
     fn set_rpc_url(&self, url: String) -> RpcResult<()>;
 
     /// Sets the base fee of the next block.
@@ -52,7 +52,7 @@ pub trait AnvilNamespaceT {
     /// # Arguments
     ///
     /// * `base_fee` - Value to be set as base fee for the next block
-    #[rpc(name = "anvil_setNextBlockBaseFeePerGas")]
+    #[method(name = "setNextBlockBaseFeePerGas", aliases = ["hardhat_setNextBlockBaseFeePerGas"])]
     fn set_next_block_base_fee_per_gas(&self, base_fee: U256) -> RpcResult<()>;
 
     /// Removes a transaction from the pool.
@@ -63,11 +63,11 @@ pub trait AnvilNamespaceT {
     ///
     /// # Returns
     /// `Some(hash)` if transaction was in the pool before being removed, `None` otherwise
-    #[rpc(name = "anvil_dropTransaction")]
+    #[method(name = "dropTransaction", aliases = ["hardhat_dropTransaction"])]
     fn drop_transaction(&self, hash: H256) -> RpcResult<Option<H256>>;
 
     /// Remove all transactions from the pool.
-    #[rpc(name = "anvil_dropAllTransactions")]
+    #[method(name = "dropAllTransactions", aliases = ["hardhat_dropAllTransactions"])]
     fn drop_all_transactions(&self) -> RpcResult<()>;
 
     /// Remove all transactions from the pool by sender address.
@@ -75,14 +75,14 @@ pub trait AnvilNamespaceT {
     /// # Arguments
     ///
     /// * `address` - Sender which transactions should be removed from the pool
-    #[rpc(name = "anvil_removePoolTransactions")]
+    #[method(name = "removePoolTransactions")]
     fn remove_pool_transactions(&self, address: Address) -> RpcResult<()>;
 
     /// Gets node's auto mining status.
     ///
     /// # Returns
     /// `true` if auto mining is enabled, `false` otherwise
-    #[rpc(name = "anvil_getAutomine")]
+    #[method(name = "getAutomine", aliases = ["hardhat_getAutomine"])]
     fn get_auto_mine(&self) -> RpcResult<bool>;
 
     /// Enables or disables, based on the single boolean argument, the automatic mining of new
@@ -91,7 +91,7 @@ pub trait AnvilNamespaceT {
     /// # Arguments
     ///
     /// * `enable` - if `true` automatic mining will be enabled, disabled otherwise
-    #[rpc(name = "anvil_setAutomine")]
+    #[method(name = "setAutomine", aliases = ["evm_setAutomine"])]
     fn set_auto_mine(&self, enable: bool) -> RpcResult<()>;
 
     /// Sets the mining behavior to interval with the given interval (seconds).
@@ -99,7 +99,7 @@ pub trait AnvilNamespaceT {
     /// # Arguments
     ///
     /// * `seconds` - Frequency of automatic block production (in seconds)
-    #[rpc(name = "anvil_setIntervalMining")]
+    #[method(name = "setIntervalMining", aliases = ["evm_setIntervalMining"])]
     fn set_interval_mining(&self, seconds: u64) -> RpcResult<()>;
 
     /// Sets the block timestamp interval. All future blocks' timestamps will
@@ -109,14 +109,14 @@ pub trait AnvilNamespaceT {
     /// # Arguments
     ///
     /// * `seconds` - The interval between two consecutive blocks (in seconds)
-    #[rpc(name = "anvil_setBlockTimestampInterval")]
+    #[method(name = "setBlockTimestampInterval")]
     fn set_block_timestamp_interval(&self, seconds: u64) -> RpcResult<()>;
 
     /// Removes the block timestamp interval if it exists.
     ///
     /// # Returns
     /// `true` if an existing interval was removed, `false` otherwise
-    #[rpc(name = "anvil_removeBlockTimestampInterval")]
+    #[method(name = "removeBlockTimestampInterval")]
     fn remove_block_timestamp_interval(&self) -> RpcResult<bool>;
 
     /// Set the minimum gas price for the node. Unsupported for ZKsync as it is only relevant for
@@ -125,7 +125,7 @@ pub trait AnvilNamespaceT {
     /// # Arguments
     ///
     /// * `gas` - The minimum gas price to be set
-    #[rpc(name = "anvil_setMinGasPrice")]
+    #[method(name = "setMinGasPrice", aliases = ["hardhat_setMinGasPrice"])]
     fn set_min_gas_price(&self, gas: U256) -> RpcResult<()>;
 
     /// Enable or disable logging.
@@ -133,7 +133,7 @@ pub trait AnvilNamespaceT {
     /// # Arguments
     ///
     /// * `enable` - if `true` logging will be enabled, disabled otherwise
-    #[rpc(name = "anvil_setLoggingEnabled")]
+    #[method(name = "setLoggingEnabled", aliases = ["hardhat_setLoggingEnabled"])]
     fn set_logging_enabled(&self, enable: bool) -> RpcResult<()>;
 
     /// Snapshot the state of the blockchain at the current block. Takes no parameters. Returns the id of the snapshot
@@ -143,7 +143,7 @@ pub trait AnvilNamespaceT {
     ///
     /// # Returns
     /// The `U64` identifier for this snapshot.
-    #[rpc(name = "anvil_snapshot")]
+    #[method(name = "snapshot", aliases = ["evm_snapshot"])]
     fn snapshot(&self) -> RpcResult<U64>;
 
     /// Revert the state of the blockchain to a previous snapshot. Takes a single parameter,
@@ -156,7 +156,7 @@ pub trait AnvilNamespaceT {
     ///
     /// # Returns
     /// `true` if a snapshot was reverted, otherwise `false`.
-    #[rpc(name = "anvil_revert")]
+    #[method(name = "revert", aliases = ["evm_revert"])]
     fn revert(&self, id: U64) -> RpcResult<bool>;
 
     /// Set the current timestamp for the node.
@@ -169,7 +169,7 @@ pub trait AnvilNamespaceT {
     ///
     /// # Returns
     /// The difference between the current timestamp and the new timestamp.
-    #[rpc(name = "anvil_setTime")]
+    #[method(name = "setTime", aliases = ["evm_setTime"])]
     fn set_time(&self, timestamp: Numeric) -> RpcResult<i128>;
 
     /// Increase the current timestamp for the node
@@ -180,7 +180,7 @@ pub trait AnvilNamespaceT {
     ///
     /// # Returns
     /// The applied time delta to the current timestamp in seconds.
-    #[rpc(name = "anvil_increaseTime")]
+    #[method(name = "increaseTime", aliases = ["evm_increaseTime"])]
     fn increase_time(&self, seconds: Numeric) -> RpcResult<u64>;
 
     /// Set timestamp for the next block. The timestamp must be in future.
@@ -188,7 +188,7 @@ pub trait AnvilNamespaceT {
     /// # Arguments
     ///
     /// * `timestamp` - The timestamp to set the time to
-    #[rpc(name = "anvil_setNextBlockTimestamp")]
+    #[method(name = "setNextBlockTimestamp", aliases = ["evm_setNextBlockTimestamp"])]
     fn set_next_block_timestamp(&self, timestamp: Numeric) -> RpcResult<()>;
 
     /// Sets auto impersonation status.
@@ -200,7 +200,7 @@ pub trait AnvilNamespaceT {
     /// # Returns
     ///
     /// A `BoxFuture` containing a `Result` representing the success of the operation.
-    #[rpc(name = "anvil_autoImpersonateAccount")]
+    #[method(name = "autoImpersonateAccount", aliases = ["hardhat_autoImpersonateAccount"])]
     fn auto_impersonate_account(&self, enabled: bool) -> RpcResult<()>;
 
     /// Sets the balance of the given address to the given balance.
@@ -213,7 +213,7 @@ pub trait AnvilNamespaceT {
     /// # Returns
     ///
     /// A `BoxFuture` containing a `Result` with a `bool` representing the success of the operation.
-    #[rpc(name = "anvil_setBalance")]
+    #[method(name = "setBalance", aliases = ["hardhat_setBalance"])]
     fn set_balance(&self, address: Address, balance: U256) -> RpcResult<bool>;
 
     /// Modifies an account's nonce by overwriting it.
@@ -226,7 +226,7 @@ pub trait AnvilNamespaceT {
     /// # Returns
     ///
     /// A `BoxFuture` containing a `Result` with a `bool` representing the success of the operation.
-    #[rpc(name = "anvil_setNonce")]
+    #[method(name = "setNonce", aliases = ["hardhat_setNonce", "evm_setAccountNonce"])]
     fn set_nonce(&self, address: Address, nonce: U256) -> RpcResult<bool>;
 
     /// Sometimes you may want to advance the latest block number of the network by a large number of blocks.
@@ -241,7 +241,7 @@ pub trait AnvilNamespaceT {
     /// # Returns
     ///
     /// A `BoxFuture` containing a `Result` with a `bool` representing the success of the operation.
-    #[rpc(name = "anvil_mine")]
+    #[method(name = "mine", aliases = ["hardhat_mine"])]
     fn anvil_mine(&self, num_blocks: Option<U64>, interval: Option<U64>) -> RpcResult<()>;
 
     /// Reset the state of the network back to a fresh forked state, or disable forking.
@@ -253,7 +253,7 @@ pub trait AnvilNamespaceT {
     /// # Returns
     ///
     /// A `BoxFuture` containing a `Result` with a `bool` representing the success of the operation.
-    #[rpc(name = "anvil_reset")]
+    #[method(name = "reset", aliases = ["hardhat_reset"])]
     fn reset_network(&self, reset_spec: Option<ResetRequest>) -> RpcResult<bool>;
 
     /// anvil-zksync allows transactions impersonating specific account and contract addresses.
@@ -268,8 +268,8 @@ pub trait AnvilNamespaceT {
     /// # Returns
     ///
     /// A `BoxFuture` containing a `Result` with a `bool` representing the success of the operation.
-    #[rpc(name = "anvil_impersonateAccount")]
-    fn impersonate_account(&self, address: Address) -> RpcResult<()>;
+    #[method(name = "impersonateAccount", aliases = ["hardhat_impersonateAccount"])]
+    fn impersonate_account(&self, address: Address) -> RpcResult<bool>;
 
     /// Use this method to stop impersonating an account after having previously used `anvil_impersonateAccount`
     /// The method returns `true` if the account was being impersonated and `false` otherwise.
@@ -281,8 +281,8 @@ pub trait AnvilNamespaceT {
     /// # Returns
     ///
     /// A `BoxFuture` containing a `Result` with a `bool` representing the success of the operation.
-    #[rpc(name = "anvil_stopImpersonatingAccount")]
-    fn stop_impersonating_account(&self, address: Address) -> RpcResult<()>;
+    #[method(name = "stopImpersonatingAccount", aliases = ["hardhat_stopImpersonatingAccount"])]
+    fn stop_impersonating_account(&self, address: Address) -> RpcResult<bool>;
 
     /// Modifies the bytecode stored at an account's address.
     ///
@@ -294,7 +294,7 @@ pub trait AnvilNamespaceT {
     /// # Returns
     ///
     /// A `BoxFuture` containing a `Result` with a `bool` representing the success of the operation.
-    #[rpc(name = "anvil_setCode")]
+    #[method(name = "setCode", aliases = ["hardhat_setCode"])]
     fn set_code(&self, address: Address, code: String) -> RpcResult<()>;
 
     /// Directly modifies the storage of a contract at a specified slot.
@@ -308,7 +308,7 @@ pub trait AnvilNamespaceT {
     /// # Returns
     ///
     /// A `BoxFuture` containing a `Result` with a `bool` representing the success of the operation.
-    #[rpc(name = "anvil_setStorageAt")]
+    #[method(name = "setStorageAt", aliases = ["hardhat_setStorageAt"])]
     fn set_storage_at(&self, address: Address, slot: U256, value: U256) -> RpcResult<bool>;
 
     /// Sets the chain id.
@@ -316,19 +316,6 @@ pub trait AnvilNamespaceT {
     /// # Arguments
     ///
     /// * `id` - The chain id to be set.
-    #[rpc(name = "anvil_setChainId")]
+    #[method(name = "setChainId")]
     fn set_chain_id(&self, id: u32) -> RpcResult<()>;
-}
-
-#[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
-pub struct DetailedTransaction {
-    #[serde(flatten)]
-    pub inner: Transaction,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[serde(default)]
-    pub output: Option<Bytes>,
-    #[serde(rename = "revertReason")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[serde(default)]
-    pub revert_reason: Option<String>,
 }
