+++ crates/api_decl/src/namespaces/mod.rs
@@ -1,0 +1,4 @@
+mod anvil;
+mod evm;
+
+pub use self::{anvil::AnvilNamespaceServer, evm::EvmNamespaceServer};
