+++ crates/api_server/src/impls/mod.rs
@@ -1,0 +1,9 @@
+mod anvil;
+mod debug;
+mod evm;
+mod net;
+mod web3;
+
+pub use anvil::AnvilNamespace;
+pub use evm::EvmNamespace;
+pub use web3::Web3Namespace;
