+++ crates/api_server/src/lib.rs
@@ -1,0 +1,4 @@
+mod error;
+mod impls;
+
+pub use impls::{AnvilNamespace, EvmNamespace, Web3Namespace};
