+++ crates/cli/Cargo.toml
@@ -13,6 +13,8 @@ categories.workspace = true
 [dependencies]
 anvil_zksync_core.workspace = true
 anvil_zksync_config.workspace = true
+anvil_zksync_api_server.workspace = true
+anvil_zksync_api_decl.workspace = true
 
 zksync_types.workspace = true
 zksync_web3_decl.workspace = true
@@ -26,6 +28,7 @@ hex.workspace = true
 itertools.workspace = true
 jsonrpc-core.workspace = true
 jsonrpc-http-server.workspace = true
+jsonrpsee.workspace = true
 futures.workspace = true
 rand.workspace = true
 serde.workspace = true
