+++ crates/core/src/node/mod.rs
@@ -1,25 +1,21 @@
 //! anvil-zksync, that supports forking other networks.
 
-mod anvil;
 mod block_producer;
 mod call_error_tracer;
 mod config_api;
-mod debug;
-mod error;
+pub mod error;
 mod eth;
-mod evm;
 mod fee_model;
-mod hardhat;
 mod impersonate;
 mod in_memory;
 mod in_memory_ext;
-mod net;
 mod pool;
 mod sealer;
 mod state;
 mod storage_logs;
+#[cfg(test)]
+mod tests;
 mod time;
-mod web3;
 mod zks;
 
 pub use self::{
