+++ crates/core/src/node/tests/mod.rs
@@ -1,0 +1,1 @@
+mod debug;
