+++ crates/types/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "anvil_zksync_types"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+keywords.workspace = true
+categories.workspace = true
+
+[dependencies]
+zksync_types.workspace = true
+
+serde.workspace = true
+
+[dev-dependencies]
+serde_json.workspace = true
